requirements.txt
Flask==3.1.1
Werkzeug==3.1.3 
Jinja2==3.1.6   
itsdangerous==2.2.0 
pandas==2.2.3
numpy==2.2.5
faiss-cpu==1.11.0
sentence-transformers==4.1.0
torch==2.7.0              
transformers==4.51.3        
huggingface-hub==0.31.2    
safetensors==0.5.3          
tokenizers==0.21.1          
google-cloud-aiplatform==1.71.1 
google-auth==2.40.1             
protobuf==5.29.4                
grpcio==1.71.0                  
requests==2.32.3
urllib3==2.4.0
charset-normalizer==3.4.2
idna==3.10
certifi==2025.4.26 